+++ desktop/cmp/clipboard/ClipboardButton.js
@@ -26,7 +26,8 @@ export const [ClipboardButton, clipboardButton] = hoistCmp.withFactory({
                 const {successMessage, getCopyText} = props;
 
                 try {
-                    await copy(getCopyText());
+                    const text = await getCopyText();
+                    await copy(text);
                     if (successMessage) {
                         XH.toast({
                             message: successMessage,
@@ -47,10 +48,11 @@ export const [ClipboardButton, clipboardButton] = hoistCmp.withFactory({
         });
     }
 });
+
 ClipboardButton.propTypes = {
     ...Button.propTypes,
 
-    /** Function returning the text to copy. */
+    /** Function returning the text to copy.  May be async. */
     getCopyText: PT.func.isRequired,
 
     /** Message to be displayed in a toast when copy is complete. */